IntervalArithmetics.jl Documentation

Contact: malek.luky@gmail.com

IntervalArithmetics Module

Functions from intervals.jl file

IntervalArithmetic.Interval — Type

Interval{T<:Number}

Struct that holds all necessary informatino about the Interval

IntervalArithmetic.Interval — Method

Interval(a::T, b::T)

Create Interval using Interval struct

IntervalArithmetic.Interval — Method

Interval(a::T)

Create Interval with only single value

Expanding Base Module functions

Functions from arithmetics.jl file

Expanded usage of Base Module

Added new input arguments with Interval struct from this Module

Comparison

Base.:== — Method

==(a::Interval, b::Interval)

Checks if the intervals a and b are equal.

Base.:!= — Method

!=(a::Interval, b::Interval)

Checks if the intervals a and b are not equal.

Base.:< — Method

<(a::Interval, b::Interval)

Compare if a < b.

Base.:> — Method

>(a::Interval, b::Interval)

Compare if a > b.

Addition and Substraction

Base.:+ — Method

+(a::Interval)

Interval does not change

Base.:- — Method

-(a::Interval)

Changes the Interval to negative values

Base.:+ — Method

+(a::Interval, b)

Addition (Interval,Number)

Base.:+ — Method

+(b, a::Interval)

Addition (Number, Interval)

Base.:- — Method

-(a::Interval, b)

Substratction (Interval,Number)

Base.:- — Method

-(b, a::Interval)

Substratction (Number,Interval)

Base.:+ — Method

+(a::Interval, b::Interval)

Addition (Interval, Interval)

Base.:- — Method

-(a::Interval, b::Interval)

Substratction (Interval, Interval)

Multiplication and division

Base.:* — Method

*(x, a::Interval)

Multiplication (Number, Interval)

Base.:* — Method

*(a::Interval, x)

Multiplication (Interval, Number)

Base.:* — Method

*(a::Interval, b::Interval)

Multiplication (Interval, Interval)

Base.:/ — Method

/(a::Interval, x)

Division (Interval, Number)

Base.inv — Method

inv(a::Interval)

Inversion (Interval)

Base.:/ — Method

/(a::Interval, b::Interval)

Division (Interval, Interval)